I started over from scratch for some reason -- I think to orthogonalize and perfect the minimal matching language.
I did this right after a realization that came after a bunch of
confusion.  I forgot that lists aren't ctons, and I was using the same
representation for both, and map took a native scheme list, not a
(cons ...).  Then I realized I needed syntax for lists, since parens
won't do it, and that made me re-think the simplified
pattern/expression representation.

I struggled with the only solution I could think of: using square
brackets for lists, which would require pre- and post-processing of
all syntax, in and out, and I didn't like that very much, but I
couldn't think of another way.  Later on I realised that (list ..)
would probably work fine, both in and out.  So I'll probably do that.

Anyway, I redid the basic pattern/exp language because I wanted to be
perfectly clear about it, and I think I did, but it might be that the
old version was fine, I don't know.  I shall keep hacking on it,
whatever.

And Erik died yesterday, by the way.  I'm hacking away whilst on the
bus to Boston to see Jeannine, to take my mind off things; note that
it's not really working.