Quote:
I'm not sure what quote is yet.  I've implemented it in patterns,
where it means "this is not a variable", but outlaw it in expressions.
This is just enough to handle 'if'.  I know this will change; after
all, quote is actually supposed to mean 'don't evaluate'.  Perhaps that's the exp-only meaning.